new update

+++ Back/src/ProEventos2.API/Controllers/EventosController.cs
@@ -1,14 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using ProEventos.Persistence;
 using ProEventos.Domain;
-using ProEventos.Persistence.Contextos;
 using ProEventos.Application.Contratos;
 using Microsoft.AspNetCore.Http;
+using ProEventos.Application.Dto;
 
 namespace ProEventos2.API.Controllers
 {
@@ -29,7 +25,7 @@ namespace ProEventos2.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if (eventos == null) return NotFound("Nenhum evento encontrado.");
+                if (eventos == null) return NoContent();
                 
                 return Ok(eventos);
 
@@ -47,9 +43,9 @@ namespace ProEventos2.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if (evento == null) return NotFound("Evento por Id não encontrado.");
+                if (evento == null) return NoContent();
                 
-                return Ok(evento);
+                return Ok(evento); 
 
             }
             catch (Exception ex)
@@ -65,7 +61,7 @@ namespace ProEventos2.API.Controllers
             try
             {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if (evento == null) return NotFound("Eventos por tema não encontrados.");
+                if (evento == null) return NoContent();
                 
                 return Ok(evento);
 
@@ -78,12 +74,12 @@ namespace ProEventos2.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.AddEventos(model);
-                if (evento == null) return BadRequest("Erro ao tentar adicionar evento.");
+                if (evento == null) return NoContent();
                 
                 return Ok(evento);
 
@@ -96,7 +92,7 @@ namespace ProEventos2.API.Controllers
         }
 
         [HttpPut("{id?}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
@@ -118,6 +114,9 @@ namespace ProEventos2.API.Controllers
         {
             try
             {
+                var evento = _eventoService.GetEventoByIdAsync(id, true);
+                if(evento is null) return NoContent();
+                
                 return await _eventoService.DeleteEvento(id) ? Ok("Deletado") : BadRequest("Evento não deletado");
             }
             catch (Exception ex)
